loop blinn
