+++ src/three/ThreeText.tsx
@@ -6,44 +6,7 @@ import type {
   TextOptions,
   ThreeTextGeometryInfo as TextGeometryInfo
 } from './index';
-
-function deepEqual(a: any, b: any): boolean {
-  if (a === b) return true;
-  if (a == null || b == null) return false;
-  if (typeof a !== 'object' || typeof b !== 'object') return false;
-
-  // Arrays (common in options like color, byCharRange, etc.)
-  if (Array.isArray(a) || Array.isArray(b)) {
-    if (!Array.isArray(a) || !Array.isArray(b)) return false;
-    if (a.length !== b.length) return false;
-    for (let i = 0; i < a.length; i++) {
-      if (!deepEqual(a[i], b[i])) return false;
-    }
-    return true;
-  }
-
-  const keysA = Object.keys(a);
-  const keysB = Object.keys(b);
-
-  if (keysA.length !== keysB.length) return false;
-
-  for (const key of keysA) {
-    if (!Object.prototype.hasOwnProperty.call(b, key)) return false;
-    if (!deepEqual(a[key], b[key])) return false;
-  }
-
-  return true;
-}
-
-function useDeepCompareMemo<T>(value: T): T {
-  const ref = useRef<T>(value);
-
-  if (!deepEqual(value, ref.current)) {
-    ref.current = value;
-  }
-
-  return ref.current;
-}
+import { useDeepCompareMemo } from '../react/utils';
 
 export interface ThreeTextProps extends Omit<TextOptions, 'text'> {
   children: string;
